frontend/requirements.txt
flask
python-dotenv
flask_wtf
email_validator
